Feature/BC-67 kcr 용 리뷰어 api 개발

+++ server/clinical-trial-system/src/main/java/com/beamworks/clinicaltrialsystem/User/Reviewer/ReviewerRepository.java
@@ -2,5 +2,5 @@ package com.beamworks.clinicaltrialsystem.User.Reviewer;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReviewerRepository extends CrudRepository<Reviewer,Long> {
+public interface ReviewerRepository extends CrudRepository<Reviewer,Integer> {
 }
